% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createBasicProject.R
\name{createBasicProject}
\alias{createBasicProject}
\title{Create a very basic project directory}
\usage{
createBasicProject(name, title = NULL, folder = getwd(),
  packagedeps = "packrat", git = TRUE, external_setup = NULL,
  reset = TRUE)
}
\arguments{
\item{name}{Project}

\item{title}{"What the Project Does (One Line, Title Case)"
If NULL, a random one will be generated.}

\item{folder}{Folder under which to create the project}

\item{packagedeps}{Set a tool for package reproducibility. Check valid options with \link{okpackagedeps}.}

\item{git}{Configure Git}

\item{external_setup}{How to do the partly interactive setup
of online git, CI and coverage platforms. If NULL, no setup.
\itemize{
\item \code{git_service} Only "GitHub" is supported at the moment.
\item \code{login} username or organization name to use for the git service.
If NULL, no GitHub repo is created.
\item \code{private} whether to make the created repo private
\item \code{protocol} "ssh" or "https", protocol to use for GitHub
\item \code{ci_activation} Only NULL, and "travis" are supported at the moment.
"travis" means calling \code{usethis::use_travis()} and adding the Travis
badge to the README.
}}

\item{reset}{whether to reset the project to current project}
}
\description{
Is the basis for all other projects.
}
\details{
├─ \strong{DESCRIPTION}          \emph{Info about authors, version and packages used (to attach them).}\cr
├─ \strong{README.md}            \emph{README for developers of this project.}\cr
├─ \strong{R}                    \emph{Source code for use in this project.}\cr
└─ \strong{<NAME>.Rproj}         \emph{Rstudio project file.}\cr
}
\examples{
\dontrun{
folder <- tempdir()
createBasicProject(name = "doggos", title = "Counting cute dogs",
                   folder = folder,
                   packagedeps = "none",
                   git = TRUE, external_setup = NULL,
                   reset = TRUE)
list.files(file.path(folder, "doggos"))
unlink(file.path(folder, "doggos"))
}
}
