% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createBatchtoolsProject.R
\name{createBatchtoolsProject}
\alias{createBatchtoolsProject}
\title{Create a project for high performance experiments with batchtools}
\source{
\url{https://mllg.github.io/batchtools/index.html}

\url{https://drivendata.github.io/cookiecutter-data-science/}
}
\usage{
createBatchtoolsProject(name, title = NULL, folder = getwd(),
  packagedeps = "packrat", git = TRUE, reset = TRUE,
  external_setup = NULL, dirs = c("data", "data/raw", "data/processed",
  "references", "reports", "reports/markdown", "reports/knitr",
  "reports/knitr/figures"))
}
\arguments{
\item{name}{Project}

\item{title}{"What the Project Does (One Line, Title Case)"
If NULL, a random one will be generated.}

\item{folder}{Folder under which to create the project}

\item{packagedeps}{Set a tool for package reproducibility. Check valid options with \link{okpackagedeps}.}

\item{git}{Configure Git}

\item{reset}{whether to reset the project to current project}

\item{external_setup}{How to do the partly interactive setup
of online git, CI and coverage platforms. If NULL, no setup.
\itemize{
\item \code{git_service} Only "GitHub" is supported at the moment.
\item \code{login} username or organization name to use for the git service.
If NULL, no GitHub repo is created.
\item \code{private} whether to make the created repo private
\item \code{protocol} "ssh" or "https", protocol to use for GitHub
\item \code{ci_activation} Only NULL, and "travis" are supported at the moment.
"travis" means calling \code{usethis::use_travis()} and adding the Travis
badge to the README.
}}

\item{dirs}{Directories to create}
}
\description{
A structure template for a project using the package \link\link{batchtools}. Inspired by Cookiecutter.
}
\details{
├─ \strong{make.R}               \emph{Commands to perform experiments (e.g. \code{\link[batchtools]{submitJobs}}).}\cr
├─ \strong{DESCRIPTION}          \emph{Info about authors, version and packages used (to attach them).}\cr
├─ \strong{README.md}            \emph{README for developers of this project.}\cr
│\cr
├─ \strong{data}\cr
│  ├─ \strong{processed}         \emph{The processed, canonical data sets for the experiments.}\cr
│  └─ \strong{raw}               \emph{The original, immutable data dump.}\cr
│\cr
├─ \strong{references}           \emph{Data dictionaries, articles, and all other explanatory materials.}\cr
│\cr
├─ \strong{reports}              \emph{Reports for project. E.g., R markdown notebooks.}\cr
│  ├─ \strong{knitr}             \emph{Generated reports as HTML, PDF, LaTeX, etc..}\cr
│  │  └─ \strong{figures}        \emph{Graphics and figures to be used in reporting.}\cr
│  └─ \strong{markdown}          \emph{R notebooks. Naming convention is a number (for ordering), the creator's initials, and a short '-' delimited description, e.g. '1.0-jqp-initial-data-exploration'.}\cr
│\cr
├─ \strong{R}                    \emph{Source code for use in this project.}\cr
│  ├─ \strong{attach-packages.R} \emph{Attach packages needed (given in description file).}\cr
│  ├─ \strong{functions.R}       \emph{Functions for experiments.}\cr
│  ├─ \strong{problems.R}        \emph{Problems for experiments.}\cr
│  ├─ \strong{algorithms.R}      \emph{Algorithms for experiments.}\cr
│  └─ \strong{design.R}          \emph{Design of experiments.}\cr
│\cr
└─ \strong{<NAME>.Rproj}         \emph{Rstudio project file.}\cr
}
\examples{
\dontrun{
folder <- tempdir()
createBatchtoolsProject(name = "doggos", title = "Counting cute dogs",
                   folder = folder,
                   packagedeps = "none",
                   git = TRUE, external_setup = NULL,
                   reset = TRUE)
}
}
