% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSwaggerProject.R
\name{createSwaggerProject}
\alias{createSwaggerProject}
\title{Create Package from Swagger API}
\usage{
createSwaggerProject(name, dir = name, url, overwrite = FALSE)
}
\arguments{
\item{name}{Project}

\item{dir}{Directory to create}

\item{url}{URL or path to swagger JSON specification}

\item{overwrite}{Should we overwrite \code{dir} if it exists? (default: FALSE)}
}
\value{
path to the newly created package path
}
\description{
Create Package from Swagger API
}
\examples{
\dontrun{
  createSwaggerProject("petstore",
      dir = getwd(),
      url = "http://petstore.swagger.io/v2/swagger.json")
}
}
