% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new.R
\name{new_project}
\alias{new_project}
\title{Create a new project}
\usage{
new_project(title = NA, short_title = NA, authors,
  current_owner = NA, creator = NA, corresp_auth = NA, stage = NA,
  deadline_type = NA, deadline = NA, id = NA,
  status = "just created", path = "", protocol = c("STROBE",
  "CONSORT"), make_directories = FALSE, use_bib = FALSE,
  datawork_template = NULL, analysis_template = NULL)
}
\arguments{
\item{title}{Title of project. Coerced to title case using
\code{tools::toTitleCase()}.}

\item{short_title}{A nickname for the project that can be used to more easily
select the project when other functions are called (e.g.,
\code{\link[=projects]{projects()}} or \code{\link{edit_project()}}).}

\item{authors}{A vector of author last_name/given_names/ids. Order will be
preserved. Each element must match a row in the \code{\link{authors()}}
table.}

\item{current_owner}{A \code{last_name}/\code{given_names}/\code{id} of one
of the authors in the \code{\link{authors()}} table.}

\item{creator}{A \code{last_name}/\code{given_names}/\code{id} of one of the
authors in the \code{\link{authors()}} table. If left blank (i.g.,
\code{NA}), it will be filled with the \code{id} of the \code{default}
author.}

\item{corresp_auth}{A \code{last_name}/\code{given_names}/\code{id} of one of
the authors in the \code{\link{authors()}} table.}
}
\description{
\code{new_project()} creates a new R project folder that is automatically
filled with a .Rproj file, helpful subdirectories, and .Rmd files to get your
project workflow started.
}
\details{
The folder and .Rproj file both have names of the form "pXXXX", where "XXXX"
denotes the project \code{id} number. The folder will be an immediate
subdirectory of the main projects folder (see \code{\link{setup_projects()}})
unless the argument \code{path} specifies a deeper subdirectory. The user may
enter various metadata about the project that is stored and may be called
forth using the \code{\link{projects()}} function. Some of this metadata will
automatically be added to a \code{\link{header}} atop the automatically
created .Rmd files called progs/01_protocol.Rmd and progs/04_report.Rmd.
}
