% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{okpackagedeps}
\alias{okpackagedeps}
\title{Valid package dependency managment}
\usage{
okpackagedeps()
}
\description{
Returns the names of options for package dependency managment in projects.
}
\details{
The options mean in detail
\itemize{
\item{none}{: No package dependency managment.}
\item{simple}{: A manual maintained file (\code{R/packages.R}) records packages needed in the project.}
\item{packrat}{: Use \link[packrat]{packrat} for package dependency managment.}
\item{checkpoint}{: Use \link[checkpoint]{checkpoint} for package dependency managment.}
}
}
\examples{
okpackagedeps()
}
