% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{projects}
\alias{projects}
\title{View the tibble of all projects.}
\usage{
projects(project, authors = FALSE)
}
\arguments{
\item{project}{An optional vector of project
\code{titles}/\code{short_titles}/\code{ids}. If each element matches a
project in the database, only those projects will be returned.}

\item{authors}{Logical indicating whether or not projects' authors will be
included in the resulting tibble. If \code{TRUE}, a left join is performed,
so that projects with multiple authors will have multiple rows in the
resulting tibble.}
}
\description{
Returns a tibble of all projects in the main projects folder, with the option
to filter by project and/or display the projects' authors as well.
}
