% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-projects_author.R
\name{vec_ptype2.projects_author}
\alias{vec_ptype2.projects_author}
\alias{vec_cast.projects_author}
\alias{projects_author-vctrs}
\title{Internal vctrs methods}
\usage{
\method{vec_ptype2}{projects_author}(x, y, ...)

\method{vec_cast}{projects_author}(x, to, ...)
}
\description{
Internal vctrs methods
}
\keyword{internal}
