% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{projects_folder}
\alias{projects_folder}
\title{projects folder path}
\usage{
projects_folder()
}
\description{
Returns the file path of the main projects folder if it has been established
via \code{\link{setup_projects}()}.
}
\details{
The file path is returned as a simple character string. It simply returns the
value of \code{\link{Sys.getenv}("PROJECTS_FOLDER_PATH")}, provided
that its value is a file path of a directory that actually exists (i.e.,
\code{\link{setup_projects}()} has been successfully run).

If it can't find a directory with that path, it returns this string:

\code{projects folder not found. Please run \link{setup_projects}()}
}
\examples{
projects_folder()
}
\seealso{
\code{\link{setup_projects}()} for setting up the projects folder.
}
