% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-projects_stage.R
\name{vec_ptype2.projects_stage}
\alias{vec_ptype2.projects_stage}
\alias{vec_cast.projects_stage}
\alias{projects_stage-vctrs}
\title{Internal vctrs methods}
\usage{
\method{vec_ptype2}{projects_stage}(x, y, ...)

\method{vec_cast}{projects_stage}(x, to, ...)
}
\description{
Internal vctrs methods
}
\keyword{internal}
