\name{toupper_first}
\alias{toupper_first}
\title{
Uppercase first letter
}
\description{
Change only the first letter of a character string to upper case
}
\usage{
toupper_first(x)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{ character string }
}
\value{
\item{character}{same as \code{x}, but with the first character in upper case}
}
\author{
Víctor Moreno-González <vmorg@unileon.es
}
\examples{
toupper_first("hola")
