% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{affiliations}
\alias{affiliations}
\alias{affilitions()}
\alias{authors}
\alias{authors()}
\alias{display_metadata}
\alias{projects}
\alias{projects()}
\title{View the \code{projects()}, \code{authors()}, and \code{affiliations()}
tables}
\usage{
affiliations(affiliation, authors = FALSE)

authors(author, affiliations = FALSE, projects = FALSE)

projects(project, authors = FALSE, archived = FALSE)
}
\arguments{
\item{projects, authors, affiliations}{Logical values indicating whether or not
to perform a left join with another metadata tibble. All \code{FALSE} by
default.}

\item{project, author, affiliation}{An optional (unique) vector of \code{id}s
and/or names. Only rows matching one or more entries will be returned.}

\item{archived}{Logical, indicating whether or not to include projects that
have been archived using \code{\link{archive_project()}} are not displayed
by default.}
}
\description{
Returns a tibble of the projects/authors/affiliations, filtered and joined
according to the entirely optional arguments.
}
\details{
If setting one or more of the \code{projects}, \code{authors}, or
\code{affiliations} arguments to \code{TRUE}, a left join is performed, with
the "left" table being the one sharing the name of the function being used.
As such, rows that don't have matches in any other tables will still show up
in the output, and rows that have multiple matches in other tables will yield
multiple rows in the output. The "right" table's id column will be renamed.

Since all these functions return \code{link[tibble]{tibble}}s, the user can
further manipulate them using \code{link[dplyr]{dplyr}} functions like
\code{link[dplyr]{select}} and \code{link[dplyr]{filter}}. See the last
example.
}
\examples{
\dontrun{
projects("Cholesterol") \%>\% dplyr::filter(current_owner == 10)

authors(affiliations = TRUE) \%>\%
  dplyr::distinct(id, .keep_all = TRUE) \%>\%
  dplyr::select(id, given_names, last_name, email, phone, address)

affiliations()
}

}
