% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_management.R
\name{file_management}
\alias{file_management}
\alias{new_project_group}
\alias{move_project}
\alias{copy_project}
\alias{archive_project}
\title{Tools for Organizing and Managing Project Files}
\usage{
new_project_group(path)

move_project(project, path = "", make_directories = FALSE)

copy_project(project_to_copy, path = "", new_id = NA,
  make_directories = FALSE)

archive_project(project)
}
\arguments{
\item{path}{A valid path string. See the \code{path} argument in
\link{new_project} for details.}

\item{project}{Project \code{id} or unambiguous substring of the project name
from the \code{\link{projects()}} tibble.}

\item{make_directories}{Logical. If the path represented by the \code{path}
parameter does not exist, should the needed directories be created?}

\item{project_to_copy}{Project \code{id} or unambiguous substring of the
project name corresponding to the project that is to be copied.}

\item{new_id}{Optional integer, ranging from 1 to 9999, used as the
newly-created project ID. Must not already exist in
\code{\link{projects()}$id}. If left blank, the lowest available `id` will
be automatically used.}
}
\description{
Projects can be moved (\code{move_project}), copied
  (\code{copy_project}), deleted (\code{\link{delete_project}}) or archived
  (\code{archive_project}). The difference between \code{delete_project} and
  \code{archive_project} is that the latter will just move the project to a
  directory called 'archive', located in the same parent directory as the
  project. This directory gets created if it doesn't yet exist.

  Projects can also be organized into groups. By default, all projects are
  created within the main \code{\link{projects_folder}}. To create a project
  group, which essentially is a subfolder of projects that sits within the
  main \code{\link{projects_folder}} (or recursively within another project
  group's folder), use \code{new_project_group()}.
}
\details{

}
\examples{
\dontrun{
# setting up a simple project directory tree
new_project_group("kidney")
new_project_group("prostate")
new_project_group("kidney/clinical")
new_project_group("kidney/genomics")
new_project_group("prostate/clinical")
new_project_group("prostate/genomics")

move_project(1, "kidney/genomics")
move_project(2, "prostate/clinical")

copy_project(2, "kidney/clinical")

archive_project(3)
}
}
\seealso{
\code{\link{new_project}}, \code{\link{delete_project}}
}
