% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup_projects}
\alias{setup_projects}
\alias{setup_projects()}
\title{Set up the projects folder}
\usage{
setup_projects(path, overwrite = FALSE, make_directories = FALSE)
}
\arguments{
\item{path}{The full file path where the user would like a directory called
"projects" to be created, wherein all \code{projects} and their data will
dwell.}

\item{overwrite}{Logical indicating whether or not to abandon any previously
stored projects folders stored in the system.}

\item{make_directories}{Logical indicating whether or not the function should
write any directories specified in the \code{path} argument that don't
already exist.}
}
\description{
Creates or restores the projects folder at the user-specified path.
}
\details{
If \code{overwrite = TRUE}, the function will run no matter what. Use with
caution.

If the user has a pre-existing projects folder and runs this command with the
pre-existing projects folder's path, nothing will be deleted.

Therefore, if the user "broke" the projects folder (e.g., by deleting
metadata; by changing the "PROJECTS_FOLDER_PATH" line in the .Renviron file),
the user can "fix" the projects folder by running this command with the
folder's actual file path.
}
